+++ src/Sulu/Bundle/ContentBundle/Search/EventListener/ReindexListener.php
@@ -125,7 +125,6 @@ class ReindexListener
                     if (!isset($count[$structureClass])) {
                         $count[$structureClass] = array(
                             'indexed' => 0,
-                            'deindexed' => 0,
                         );
                     }
 
@@ -133,13 +132,8 @@ class ReindexListener
                         continue;
                     }
 
-                    if ($structure->getNodeState() === Structure::STATE_PUBLISHED) {
-                        $this->searchManager->index($structure, $locale);
-                        $count[$structureClass]['indexed']++;
-                    } else {
-                        $this->searchManager->deindex($structure, $locale);
-                        $count[$structureClass]['deindexed']++;
-                    }
+                    $this->searchManager->index($structure, $locale);
+                    $count[$structureClass]['indexed']++;
                 } catch (\Exception $e) {
                     $output->writeln(
                         '  [!] <error>Error indexing or de-indexing page (path: ' . $node->getPath() .
@@ -154,15 +148,14 @@ class ReindexListener
         $output->writeln('');
 
         foreach ($count as $className => $stats) {
-            if ($stats['indexed'] == 0 && $stats['deindexed'] == 0) {
+            if ($stats['indexed'] == 0) {
                 continue;
             }
 
             $output->writeln(sprintf(
-                '<comment>Content</comment>: %s <info>%s</info> indexed, <info>%s</info> deindexed',
+                '<comment>Content</comment>: %s <info>%s</info> indexed',
                 $className,
-                $stats['indexed'],
-                $stats['deindexed']
+                $stats['indexed']
             ));
         }
     }